requirements.txt
selenium>=4.16.0
Pillow>=10.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.1
pyinstaller>=6.3.0
